Cull offscreen tiles
Right now, this'll cull the tiles by removing and spawning the sprites when needed.
In the future, the sprites will simply be recycled to save time.
The culling positions are hardcoded currently,
so maps that don't use 40x40 tiles (or 20x20 2x) will look odd.

This does not fix the collision being offset.